agentic-ai-platform/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.10.0
tiktoken>=0.5.1
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.5